Hb: Formalize Mustache "names" in PSI tree
Formally wrap the part of a mustache which is it's "name" to clean up
a number of lookups which rely on this (i.e. auto-insert of close tags
needs to grab the name to write the correct closer).

This also makes the structure of mustache names analogous to the
structure of partial names in the PSI tree.