WEB-15350 completion popup show private variables
https://youtrack.jetbrains.com/issue/WEB-15350

I've tried to fix it like this. I only show private variables if the user begins his search with `_`. see DartReferenceCompletionContributor.java.

And if the user does indeed begins his search with _ this, I try to make sure only variables that are part of the library are shown. See DartResolveUtil.java