Made check for "addEventListener" method precise.
ActionScriptTypeChecker complains if a method "addEventListener" is
called callback that does not have a single parameter of type flash.events.Event.
However, the check used to simply compare the method name, not whether it
actually refers to flash.events.IEventDispatcher#addEventListener(). Thus, if a method
in another framework is named "addEventLister", you get false negatives.

This fix checks whether the object on which addEventListener() is invoked is of
a type that implements Flash's IEventDispatcher interface.