Test for fix of IDEA-98836
Set up test source with a properties file is located in a package folder "resourcepackage"
and an MXML file using this resource bundle.
Check that it can no longer be referenced by its simple name "mybundle4", but only by its
fully-qualified name "resourcepackage.mybundle4".
Check that the key "prompt5" is actually found (not an error).

The existing tests #testResourceReferences() and #testResourceReferences2() ensure
that for the usual top-level package properties files, everything still works as before.