TSLint language service: parsing of the linter result is done on IDE side now. From performance point of view, it is better: we are parsing language service answer anyway
(On the language service side - we are getting the result of linting as string and putting into JSON object. Makes no sense to either parse-serialize or construct result output from string pieces, how it was before.
Now we get additional error field, which can (and contains) symbols that should be escaped on serialization, and the best way to do this is JSON.stringify, so at least part of result object needs to be constructed with help of JSON.stringify)

WEB-26046 Tslint not working on Webstorm 2017.1