improve notifications in case eslint/tslint service did not start
- do not write into event log, use editor notifications
- if language service getProcess() returns null === service was not
started because of error => return FixedFuture with error to the client
- when waiting for language service response, do not ignore
Future.get() exceptions - they are thrown when the execution pool of the
service is being stopped - treat as cancel
- when waiting for language service, additionally check that language
service was not stopped: in case of the first run of linter annotator,
at the moment, when the linter task was scheduled,
there is no information about the service state (it will be started),
the task is scheduled.
But before its execution starts, the pool is cancelled.
That's why without the check the task would wait for timeout.
- when getting language service error, also check for the error from
protocol (it can contain information also if process was started)
- logging of language service start error should not be ERROR level,
it depends on users settings
WEB-27418 TSLint shows error message when opening new project with tslint.json and ? no installed TSLint?
WEB-26426 Improve notifications and error messages from TSLint

r=Starovoyt
(cherry picked from commit 4b869b6)

(cherry picked from commit 879758f)