SQL formatting: separate code style constants.
Constants from SqlCodeStyleSettings are moved into a new 'interface' SqlCodeStyleConst. Both variables and constants in the original files are written in upper case (I don't know what for), and it's not ease to distinguish them. So when they are in different files they look better.
In order to preserve compatibility with possible plugins, SqlCodeStyleSettings implements SqlCodeStyleConst.