Add request listener methods to AnalysisServer interface.
This patch builds on the work in d6a6512b6ea0f1faee3cebd6c05271bed9369b99
which added a mechanism to subscribe to notifications whenever the Dart
analysis server client receives a response. Here we're adding the
symmetric capability for subscribers to be notified whenever a request
will be issued. This new hook allows tools interfacing with the Dart
plugin (like Flutter) to do things like measure roundtrip times to the
Dart analysis server.