PY-22549: Support different test duration calculation strategies.
By default, test durations are set explicitly and suites are using sum of durations of their children.

This is ok for serial tests, but does not work well for parallel testing.

This change adds ability to provide TestDurationStrategy (see appropriate class) for each suite including root. This strategy is then used by SMTestProxy