Issue 35680 - Dart code completion suggestions from not yet imported libs
This change is associated with the analysis server protocol changes from
https://github.com/dart-lang/sdk/commit/d4b8e0696df14a7834b2f89659b7e3fc41fd3c91
which @scheglov has added support for on the server side. At a high
level, the flow has client send server a subscription for the available
suggestion sets. Server will then send availableSuggestions notifications
that give the client pending completions to cache. At completion time,
we include a couple new fields in CompletionSuggestion which allow us to
identify which of the cached suggestions should be included (in addition
to the default ones).