[IFT] do not override actions from RecentProjectListActionProvider (IFT-12)
Previously we had to override RecentProjectListActionProvider to add Tutorials actions for it. RecentProjectsPanel operates with this list of actions and any list update (removing ProjectGroup for example) putting actions from the provider back. In this commit we are adding all the custom actions (Tutorials) for the ActionList when the model has been cleared. It gives us a way not to customize actions provider at all.

GitOrigin-RevId: 6f4f66368b06dd8c5c262b48d45dcb49856408a3