move static methods from CommonBundle to AbstractBundle
CommonBundle provides access to CommonBundle.properties, and in order to fix IDEA-203640 it should be moved out of util module, this will make it inaccessible from some modules which used static methods from it, moving them to AbstractBundle will help to avoid this. Also it's more natural to have generic utility methods in AbstractBundle which isn't bound to a specific *.properties file.

GitOrigin-RevId: 66a1db7ac932565ff8e37f14e3a3711ff17e2ea5