Update to bnd 5.0.0
bnd 5.0 was recently released and some of our customers are asking for the updated OSMORC plugin. This patch is unfortunately not tested because I am not sure how to build and test this repository? Is it possible to streamline these updates? Just create a PR like this? Could we build a snapshot version of the plugin in our build?

There seem to be a number of companies interested in this plugin and it would be nice if we could work together to keep this plugin synchronized with the bnd release cycle.