[flex tests] don't register SDK returned from LightProjectDescriptor
This JDK is registered in LightProjectDescriptor.registerSdk, so before this change it was added to ProjectJdkTable twice. And during tearDown when removeJdk was called for a first time, it disposed ProjectJdkImpl instance and fired jdkRemoved event, and RootIndex tries to rebuild itself and accessed content of disposed ProjectJdkImpl instance (because the second instance is still present in ProjectJdkTable).

GitOrigin-RevId: 3fb2bce0a00aa15ce36a0441e7d52b706dbfb067