IDEA-CR-65928: [ruby] Do not change gem load process for minitest 5.0 and newer.
In modern version of minitest it's possible to add the reporter and make monkey patch via file "minitest/*_plugin.rb". So it's possible not to break the load process.

For older minitest versions (the latest version of 4.7 was released in 2013) we still have to load rubymine_test_framework_initializer.rb before the main application. But it's only ~ 1% of our users.

RUBY-26848 2020.2 uses minitest_runner.rb for running/debugging individual tests and it's causing problems (at_exit handler of a gem runs before the test)

GitOrigin-RevId: 2f6db8a981b3155a70e2ac11efe2964ee8b7cb28