[jps model] cleanup: remove code which implements save for JPS model
JpsModel is used inside the build process (except small part related to custom source root types which is also used inside IDE) and the build process doesn't save the project configuration. And we don't have plans to use JpsModel inside the IDE process anymore (the new workspace model is used instead), so it makes no sense to maintain that code.

GitOrigin-RevId: 9cfa7ccc86756550ccf8d54d03e3e267fdd5933f