[IFT] Edit the way of creating Python learning project
In PyCharm it will be created using logic from python new project wizard (this is not very convenient implementation, since we have to create UI components where they are not needed). In another IDE with Python plugin installed it will be created as before.
These changes were made in order to choose the most convenient SDK and reduce the indexing time.

IDEA-CR-68001

GitOrigin-RevId: c91e0f5c25dfad1bc98cb6840f263c4a40136806