[IFT] Edit the way of choosing JDK for Java learning project
Earlier if we couldn't find local JDK, it chooses the JDK that bundled in IDEA (JBR). But it is forbidden to use JBR outside of IDEA.
Now JDK lookup is done using SdkLookup. It will try to find the local JDK, if not, it will prompt the user to download the latest version of JDK.

IDEA-CR-68001

GitOrigin-RevId: ac771bdee47a2503aa2e2febe9cc7c68eb093052