[workspace model] support SdkFinder extension to fix problems in Flex apps (IDEA-256952)
FlexCompositeSdk instances were created via SdkFinder extension in the old project model, so we need to support it in the new model as well to avoid exceptions in Project Structure dialog for Flex projects. This change also stores created FlexCompositeSdk instances in an application service to avoid creating new instances and registering their listeners each time JdkOrderEntry::getJdk is called.

GitOrigin-RevId: 3c7867d498a54e0fb61915fa5e4c65f8822c38a6