clion: try to minimize use of io.File in toolchain-related code
The File instances are primarily fake files, they do not exist in
the real life as a toolchain can be remote for example. They also
cause issues when path separators are changed to Windows ones
for unix paths and do not look nice in UI. The only value of File
use is an option to get file name, which is now extracted to
CidrExecutableTool.

GitOrigin-RevId: c178152758b962c51a02b73b340ec3d1d6f9e27b