[workspace file index] rewrite RootsChangeWatcher to work with WorkspaceFileIndex, not WorkspaceModel
Before RootsChangeWatcher searches for roots in entities of Workspace Model. It wasn't quite correct, because there may be a file referenced from an entity which shouldn't trigger rootsChanged event, and WorkspaceFileIndexContributor may register a root which isn't present in an entity (e.g. to exclude a subdirectory of a directory mentioned in an entity). Now this functionality is moved to NonExistingWorkspaceRootsRegistry, and it works with roots registered in WorkspaceFileIndex. Implementations of WorkspaceFileIndexContributor should pass VirtualFileUrl instances instead of VirtualFile to ensure that data in the index is properly updated if the file is created later.

The old implementation of RootsChangeWatcher is kept under OldRootsChangeWatcher name for the cases when WorkspaceFileIndex is switched off.

GitOrigin-RevId: 6e98f8345cb0ef8816f27d522235ed1a0187eb3d