LSP work in progress. Quite a refactoring which I failed to do in small chunks, sorry. Details below.
LspServer class is the main entity that represents each started LSP server. LspServerManager manages the list of LspServers. LspServerDescriptor still provides info, how to start the server and tunes server behavior, but it's not a key in any map anymore and is not a facade for server methods.

Unsaved files management fixed (WEB-55496). File contents synchronization (didOpen/didChange/didClose) now used for all files which are open in an editor and also for all other unsaved files.

"Slow operation" assertions fixed (IDEA-302321).

Lots of related cleanup. And more to come.

GitOrigin-RevId: cd8c96b6728105cf4464e31799b72378efb41df9