[protoeditor] IDEA-317580 Added an option to search for proto files in indexes during resolve
This option is considered to be a fallback for cases when an import statement can not be resolved with the manually configured import paths set, but user doe not want to configure it porecisely. Thus we can search for proto files in the file type index and check whether their corresponding import statement matches found file url. Having the option enabled, users would have less control over the imported files configuration, but more freedom not to think about IDE configuration and to do the actual work instead

GitOrigin-RevId: 75964fea181ff412600cb08aaab510833563f129