[dart] don't access WorkspaceFileIndex before the project is initialized (EA-815361)
WorkspaceFileIndex is initialized while the project is opening, and it shouldn't be accessed before the project is initialized. It seems that the only place where DartFileListener.scheduleDartPackageRootsUpdate can be called on an uninitialized project is DartWorkspaceModelChangeListener, and now it'll check for that (it'll be called from DartStartupActivity in such cases). And an assertion is added to DartFileListener just in case.

GitOrigin-RevId: 22ec241b6cce32c51e09e9f8462ed0277860f599