IJPL-238 update OSGI manifest assertions and add assertJ library for tests
Updated the OSGI manifest assertions for more readability in the code. This was done by rearranging set elements to separate lines. Additionally, the "java.lang" package was added to the "import-package" for each manifest. This change may help in debugging and improving code readability.

Additionally, assertJ library was added to the test scope. This was done as a replacement of assertEquals function to simplify assertion in OsgiBuildTestCase.kt file and provide more readable assertion error messages.

GitOrigin-RevId: 8506b4855efa8fbfe422f0d5de2f717d6044d87c