[Karma Debugger] WEB-63473 Add the ability to debug Chrome and Chromium browsers in Karma runner
Previously, it was possible to debug only specially configured browsers (ChromeHeadless, ChromeCanaryHeadless, ChromiumHeadless) or browsers from the customLaunchers block with the --remote-debugging-port flag. This resulted in the user being unable to debug configurations that, for example, only use Chrome.

This will allow to debug such configurations due to the additional configuring browsers for debugging. We move such browsers to the customLaunchers block and add the --remote-debugging-port flag to them.


Merge-request: IJ-MR-117384
Merged-by: Shumaf Lovpache <Shumaf.Lovpache@jetbrains.com>

GitOrigin-RevId: 87c58e462e2c6e40ddcce9c11e0889afb807c581