[terraform] IDEA-348422 IDEA-348734 Upgraded lexer and BNF to ensure sustainability of the parsed tree
The new lexer works well with incremental updates because it makes a decision whether a particular segment is a data language segment or a template one in a separate state. Thus, initial state now truly means `no info about context` instead of 'we are in a data lang segment now' as before.

Updated parser handles erroneous elements well too, because it is now capable of processing BadTags wherever they might occur in an incomplete file

GitOrigin-RevId: bb6ceb5c36aeb8f0259007fab4c4876e06062b57