Add basic editions support to intellij protobuf plugin.
This updates the protobuf grammar to include the following changes in edition 2023:
- Addition of `edition` keyword
- Reserved fields and enum values are now specified as identifiers instead of string literals

This also adds some basic validation checks:
- Enforce that only supported editions are allowed
- Enforce reserved fields and enum values obey the file's syntax
- Prohibit the removed `optional`, `required`, and `group` keywords
- Prohibit the `features` option from being used outside editions

As an immediate follow up we will audit all the gencode -> proto utilities exposed by the plugin and at least add some editions tests.

Future changes could actually handle feature cascading and provide some of the additional checks that protoc will apply during builds.  For example, every feature specifies the type of descriptors it's allowed to be specified on, and certain combinations of features are prohibited.