[Dart] IDEA-349057 follow-up on Dart Macros support: disable Reader Mode
Auto-formatting applied in Reader Mode for Dart macro-generated files and libraries may look confusing. One of the reasons is that the internal formatter is not the canonical source of truth, and the 'dart format' tool is not invoked in Reader Mode. It's better to see Dart files as they are.

GitOrigin-RevId: 39f01dfcf929d34c18209c147cc8a57d29bbea54