[qodana] correctly add YAML to classpath when it's extracted to a separate module
When intellij.yaml.editing module is extracted to a separate JAR (this is needed for IJPL-176102), it will be put to a separate JAR, so we need to add it explicitly. Also, in that case, it may be loaded by a separate classloader, which won't have plugin.xml, so it should be possible to add 'pluginAdditionalJars' even if the main JAR isn't found.

GitOrigin-RevId: 67e102d7a5ee165a306621009087aedd2ecaf6dd