RUBY-31499: send messages about started/failed tests in the Cucumber 4+ formatter.
The messages do get sent in the older formatters but not in the Cucumber 4 one.

Additionally, this commit removes the warning icon that gets displayed for tests with custom categories if the progress text was not empty, since otherwise there was a warning icon for all Cucumber tests. The line setting the icon was added for tests with incomplete indices, but the warning for that is also properly set elsewhere, so it appears the warning is not necessary.

IJ-CR-156354

(cherry picked from commit 44a08affc1849b64fe2345da4265231fb735005c)

GitOrigin-RevId: 453e3f81dbc915d99cafe7ec1809256903bb3cbc