[external-systems] Use ProcessOutputType instead of boolean in ExternalSystemTaskNotificationListener.onTaskOutput
Using a boolean stdOut parameter makes the external systems framework lose information about the SYSTEM process output type.

In particular, it leads to issues like AMPER-4430 and KMT-950 when the system output is treated either as stdOut or as stdErr.

It is a breaking change. Unfortunately, there is no way to keep compatibility, because ExternalSystemTaskNotificationListener is both extended and called by plugins.

GitOrigin-RevId: dac8809009f9226ccceff7d4235b102cb3dafd39