[js debugger] WEB-74119 Finish migration to Plugin Model V2
All code was moved from the root module to the `backend`. This allows to remove hacks with `loading="embedded"` and inclusion into `javascript.frontend.split` because now all the modules in the plugin are loaded based modules available in the environment (client/monolith mode).

Also `intellij.javascript.debugger.javaee.jsp` module was introduced to fix IJPL-165543 and some issues with loading on the client.


Merge-request: IJ-MR-170868
Merged-by: Shumaf Lovpache <soarex16@gmail.com>

GitOrigin-RevId: 63ef497a4289b2b23a4eb2e7a85b6e81803cb6e9