[platform] convert intellij.platform.scriptDebugger.ui to content module (IJPL-200937)
Explicit dependency on it is added to plugins which use its classes. To keep compatibility with external plugins, the module is added to 'contentModulesExtractedInCorePluginWhichCanBeUsedFromExternalPlugins' list. Also, a plugin alias is introduced to allow adding explicit dependency on the module via 'depends' tag.

'intellij.platform.debugger' and 'intellij.platform.debugger.impl' are added to app.jar explicitly in PlatformModules (before they came as dependencies of intellij.platform.scriptDebugger.ui or intellij.platform.debugger.parallelStacks, but the latter isn't included in all products).

GitOrigin-RevId: 381b1e817af61ca360b4c0ec6816d6167fba26a2