[clion, platformio] CPP-30672 Better project reload logic
Refactored the reload-on-run-config-change logic to better reflect when we actually want to reload. The only case when the reload is needed and not already done by other listeners is when we change from platformio RC to another platformio RC and the environment variables have changed. This covers both changing from one RC to another and changing the env variables in the selected RC. To deduplicate multiple reloads caused by multiple edits in one edit dialog, we collect the changes only in the non-modal state.

GitOrigin-RevId: 2673bc10af0b5b08de00315e52087dd1344c8865