[cucumber-java] JavaAnnotatedStepDefinition: point to PsiAnnotation, not PsiMethod
This turned out to be a much more widespread change than I expected it to be,
mainly because the code in this plugin seems to be all over the place.

I tried to find every place that assumed that JavaAnnotatedStepDefinition
is always backed by PsiMethod and changed this assumption to PsiAnnotation.
All tests pass.

GitOrigin-RevId: 622b8971a1d43a9f50110d81439d11843ee27722