[js debugger] IJPL-197290 Add RemoteBrowserSelector
Currently, all run configuration editors execute UI on the backend (BE Controls), so BrowserSelector uses a list of browsers from the host. As a result, when a user opens a run configuration editor, he sees browsers on a remote machine, not his own.

This problem can be solved in the following ways:

- Reimplement run configuration editors with RPC. This solution is technically correct but very difficult to implement and therefore will not be implemented in the near future.
- Execute a specific UI component on the client - this is possible, but not in modal dialogs.
- Implement a component that will be executed on the backend but request information from the frontend via RPC. Although this solution looks like a crutch, it can be implemented pointwise (in the browser debug run configuration) and it does not require large implementation costs.

This commit implements the latter approach introducing `RemoteBrowserSelector` and adjusting some API's to use abstract `browserId` instead of the `WebBrowser` which represents browser on the local machine.


Merge-request: IJ-MR-173835
Merged-by: Shumaf Lovpache <Shumaf.Lovpache@jetbrains.com>

GitOrigin-RevId: 3f560d55c03ba7941927a360235e5e324f852eea