[cucumber, python] remove inplace rename for Gherkin step parameters
This change is aimed at reducing the number of moving
things to help debug IDEA-376182.

It might be added back later on, with more tests.

---

Test `PyBDDRenameTest#testNoDefinitionNotAllowed` started to fail because
its test fixture contains `<warning>` tags, which are interpreted as `GherkinStepParameter`s,
causing the default rename handler in `RenameHandlerRegistry#doGetRenameHandlers` to be used.

The simplest fix was to move such a test to CucumberJavaStepRenameTest,
whose test fixture doesn't contain such `<warning>` tags.

GitOrigin-RevId: 90e636d969f4394fa5439c5de6a3e5a199e218e9