[serial-monitor] CPP-46597 Fix corrupted state with multiple opened windows
In the settings, update profile logic to also prefer state of the opened tab for the logic behind showing/enabling the connect/disconnect/reconnect buttons. Similar logic is implemented for ports.
Now when a port is opened from a different window, only the reconnect button is visible for both profiles and ports, which disconnects the connection in the other window (does not close the tab) and then connects in the current window.

GitOrigin-RevId: 97d8f3be4e69b0c4bbd7bfab599354a1c274ebf0