[cucumber] IDEA-376182 fix renaming step parameters working only when the step is unresolved
### Summary of changes

Introduced a new GherkinStepParameterRenameHandler.
It is used when renaming step parameters.

GherkinStepParameter is no longer a PsiNameIdentifierOwner (which itself is a PsiNamedElement).
GherkinStepParameter is a usage, not a definition. It references GherkinTableCell.
And Javadoc of PsiNamedElement say:
"References should not implement this interface since they don't usually introduce a new entity."
It didn't make sense for GherkinStepParameter to be a PsiNamedElemet.

At first, I thought that making GherkinStepParameter not be a PsiNamedElement would fix the issue, but it didn't.
Still, a nice cleanup.

### More detailed description

This was a very strange problem.
Even the title of the bug might cause you to think "what?!".

I started investigation by looking into CodeInsightTestFixtureImpl.renameElementAtCaretUsingHandler.
It calls EditorUtil.getEditorDataContext(editor) and the DataContext returned by it differs depending on
whether the step (inside which the step parameter was) was resolved or unresolved.

Case 1: the step is unresolved

  Calling getData(CommonDataKeys.PSI_ELEMENT) on that DataContext returns a GherkinTableCell.
  This is even weirder because it happens no matter if rename is called on the element defining the step parameter (i.e. GherkinTableCell),
  or on the element referencing the step parameter (i.e. GherkinStepParameter).
  In both these sub-cases, DataContext#getData(CommonDataKeys.PSI_ELEMENT) returns a GherkinTableCell.

Case 2: the step is resolve

  Calling getData(CommonDataKeys.PSI_ELEMENT) on that DataContext returns null.

This still happens and is reproducible.
Put a breakpoint in CodeInsightTestFixtureImpl.renameElementAtCaretUsingHandler
and debug tests inside CucumberJavaStepParameterRenameTest.

I don't know *why* it happens, but at least it's fixed now (thanks to GherkinStepParameterRenameHandler).
If you ever find out the cause of this behavior, please share your findings in some visible place.
and I also wrote tests that verify that renaming a step parameter works in different cases:
- rename definition (GherkinTableCell), step is resolved
- rename usage (GherkinStepParameter), step is unresolved
- rename usage (GherkinStepParameter), step is resolved

GitOrigin-RevId: 8a0c1508ed68a11b89fa1331a1a777e9617b2206