[cucumber-java] IDEA-202256 fix occasional error "unexpected start of quantifier '{'" in step definitions
The title of the ticket is misleading.

Sometimes, the version of the cucumber-jvm library, present in the project, can't be determined.
In such cases, CucumberJavaVersionUtil.computeCucumberCoreVersion should default to the latest version, not the first one, i.e. v1 (which is ancient by now and almost no-one uses it).

CucumberJavaInjector uses resul of CucumberJavaVersionUtil.computeCucumberCoreVersion to decide
whether to inject regex language into Given/When/Then annotations.
This is how the bug came to be.

GitOrigin-RevId: aae2178dae07a0074b238dfcd1ee5ba373524b18