[cucumber] IDEA-99170 CucumberStepInspection: warn about ambiguous step references
Fix test failures in PyBDDStepResolveTest which turned out to be a bit problematic.
It took me some time to fix the failing test there: testAndKeywordResolvesCorrectly.
It all came down to the tags like <warning descr="..."> not being removed until
CodeInsightTestFixture.testHiglighting is called.
And also, because of that, moving to the caret was broken (because the warning tags affected the number of characters in the file).

And also the "Create step definition" is not present on the whole GherkinStep,
but only on its part as determined by CucumberStepReference#getRangeInElement.

Death by a thousand cuts.


Merge-request: IJ-MR-177276
Merged-by: Bartek Pacia <bartek.pacia@jetbrains.com>

GitOrigin-RevId: 42cdb337bd3354e2676152da25939fcdb4459cfb